# Compress operator parsing hierarchy