This groups operators that work on different types into a single precedence
level to improve parsing performance.  Parsing the precedence of these groups
of operators incorrectly doesn't matter since any expression that mixes them is
ill-typed

This improves performance by about 15% on the benchmark expression from #108

Before:

```
time                 334.4 ms   (308.6 ms .. 375.3 ms)
                     0.998 R²   (0.995 R² .. 1.000 R²)
mean                 349.8 ms   (340.4 ms .. 364.3 ms)
std dev              14.35 ms   (5.761 ms .. 19.71 ms)
variance introduced by outliers: 19% (moderately inflated)
```

After:

```
time                 289.9 ms   (274.6 ms .. 297.9 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 309.3 ms   (301.0 ms .. 318.3 ms)
std dev              11.08 ms   (8.083 ms .. 14.31 ms)
```