@Gabriel439 2018-09-26
Note that I also tried using `Text.Parser.Expression.buildOperatorExpression` as well and that made things slower (even slower than the original code)