@Gabriel439 2018-09-26
This also improves heap residency, which goes down from ~5.5 MB to ~4 MB